+++ web/src/components/Panel.tsx
