+++ web/src/features/auth/components/UserDropDownMenu.tsx
@@ -14,12 +14,12 @@ import {
 } from "@chakra-ui/react";
 import React, { ReactElement } from "react";
 import { FiChevronDown } from "react-icons/fi";
-import Link from "../../../components/link";
+import Link from "../../../components/Link";
 import {
   useLoginState,
   useLogout,
   UserAttributes,
-} from "../../../hooks/use-authentication";
+} from "../../../hooks/useAuthentication";
 
 export function LoginControl(): ReactElement {
   return (
