+++ web/src/features/auth/routes/Login.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import { useNavigate } from "react-router-dom";
-import { useLoginState } from "../../../hooks/use-authentication";
+import { useLoginState } from "../../../hooks/useAuthentication";
 import { LoginForm } from "../components/LoginForm";
 
 export const Login = (): ReactElement => {
