+++ web/src/features/creators/components/ShowCreator.tsx
@@ -1,6 +1,6 @@
 import { Box, Heading, List, ListItem, Stack, Text } from "@chakra-ui/react";
 import React from "react";
-import Link from "../../../components/link";
+import Link from "../../../components/Link";
 import { Creator } from "../../../types";
 
 interface Props {
