+++ web/src/features/publishers/components/ViewPublisher.tsx
@@ -1,6 +1,6 @@
 import { Box, Heading, List, ListItem, Stack, Text } from "@chakra-ui/react";
 import React from "react";
-import Link from "../../../components/link";
+import Link from "../../../components/Link";
 import { Publisher } from "../../../types";
 
 interface ViewPublisherProps {
